Second full code-review pass, all device-tested on the Miyoo Flip:
- Batch 1 (+R2/R3): sample-edit locking, SF eviction races, modal guards, RATE pitch
- Batch 2 (+R2): int64 START/END, streaming WAV export, load-OOM, stereo preview
  markers, song-selection scroll
- Batch 3: low-risk perf/cleanup (logging gates, hex/font caches, FFT cfg cache)
- Batch 4 + 4b: eliminate idle 60fps full-screen redraw; OCTA preview lane
- Batch 5: audio hot-path 1.2/1.3/1.10 (per-block queue drain, trimmed per-callback
  zeroing, demand-gated spectrum capture)

Remaining Review #2 items (smaller, deferred): 6.2 glyph atlas, 7.2 R8 minify,
7.3 native -O2, 1.8/4.1/3.3/2.4 robustness/cleanup.

+++ app/src/main/cpp/sampler-voice.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <cmath>
+#include <cstdint>
 #include <cstdlib>
 #include <algorithm>
 #include "mods/mod-system.h"
@@ -120,9 +121,11 @@ struct Voice : public IAudioVoice {
         // Use startPointOverride if provided (Offset effect / slice start), otherwise use instrument default
         int effectiveStartPoint = (startPointOverride >= 0) ? startPointOverride : instrParams.startPoint;
         int effectiveEndPoint   = (endPointOverride   >= 0) ? endPointOverride   : instrParams.endPoint;
-        actualStart = (effectiveStartPoint * length) / 255;
-        actualEnd   = (effectiveEndPoint   * length) / 255;
-        actualLoopStart = (instrParams.loopStart * length) / 255;
+        // int64 math: point × length overflows int32 for samples ≳ 8.4M frames (~3 min at
+        // 44.1 kHz — easy to hit via video-audio extraction), silently breaking START/END.
+        actualStart = (int)(((int64_t)effectiveStartPoint * length) / 255);
+        actualEnd   = (int)(((int64_t)effectiveEndPoint   * length) / 255);
+        actualLoopStart = (int)(((int64_t)instrParams.loopStart * length) / 255);
 
         // Clamp to valid range
         actualStart = std::max(0, std::min(actualStart, length - 1));
@@ -307,7 +310,8 @@ struct Voice : public IAudioVoice {
     void retrigger(int startPoint) override {
         if (!isActive || !sampleData) return;
         if (startPoint >= 0 && startPoint <= 255 && sampleLength > 0) {
-            position = (float)((startPoint * sampleLength) / 255);
+            // int64 — same overflow as trigger() for long samples
+            position = (float)(((int64_t)startPoint * sampleLength) / 255);
             position = fmaxf((float)actualStart, fminf(position, (float)(actualEnd - 1)));
         } else {
             position = (float)actualStart;
